Added remote footer functionality
The remote footer support is a workaround for the fact that
ONA still isn't capable of handling external references
in terms of DNS records.

Remote footers can be enabled by using the option "-t".
For each domain on the name server the build_bind script
will check for an available footer file on a web server.
Instructions are included with the README file.